requirements.txt
Flask==3.0.2
fpdf2==2.7.8
